Add 'rpc' feature instead of 'test' and 'execute'
