+++ tooling/nargo/src/lib.rs
@@ -9,15 +9,12 @@
 
 pub mod constants;
 pub mod errors;
+pub mod foreign_calls;
 pub mod ops;
 pub mod package;
 pub mod workspace;
 
-#[cfg(feature = "execute")]
-pub mod foreign_calls;
-
 pub use self::errors::NargoError;
-#[cfg(feature = "execute")]
 pub use self::foreign_calls::print::PrintOutput;
 
 use std::{
