+++ tooling/nargo/src/ops/mod.rs
@@ -6,17 +6,12 @@ pub use self::compile::{
 pub use self::optimize::{optimize_contract, optimize_program};
 pub use self::transform::{transform_contract, transform_program};
 
-#[cfg(feature = "execute")]
 pub use self::execute::{execute_program, execute_program_with_profiling};
-#[cfg(feature = "test")]
 pub use self::test::{run_test, TestStatus};
 
 mod check;
 mod compile;
-mod optimize;
-mod transform;
-
-#[cfg(feature = "execute")]
 mod execute;
-#[cfg(feature = "test")]
+mod optimize;
 mod test;
+mod transform;
